backend/requirements.txt
# FastAPI and web server
fastapi==0.103.1
uvicorn==0.23.2
gunicorn==21.2.0
pydantic==2.3.0
starlette==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
asyncpg==0.28.0

# Authentication and security
passlib==1.7.4
pyjwt==2.8.0
bcrypt==4.0.1
python-jose==3.3.0
cryptography==41.0.3
# itsdangerous==2.2.0

# Utilities
python-dotenv==1.0.0
email-validator==2.0.0
httpx==0.25.0
requests==2.31.0
tenacity==8.2.3
setuptools==68.1.2
ujson==5.8.0 